GS, UI: use u32 instead of s32 for width and height

+++ pcsx2/GS/Renderers/Metal/GSDeviceMTL.mm
@@ -1267,11 +1267,11 @@ std::string GSDeviceMTL::GetDriverInfo() const
 	return desc;
 }}
 
-void GSDeviceMTL::ResizeWindow(s32 new_window_width, s32 new_window_height, float new_window_scale)
+void GSDeviceMTL::ResizeWindow(u32 new_window_width, u32 new_window_height, float new_window_scale)
 {
 	m_window_info.surface_scale = new_window_scale;
 	if (!m_layer ||
-		(m_window_info.surface_width == static_cast<u32>(new_window_width) && m_window_info.surface_height == static_cast<u32>(new_window_height)))
+		(m_window_info.surface_width == new_window_width && m_window_info.surface_height == new_window_height))
 	{
 		return;
 	}
